Refactored away YouTubeEmbed component
